refactor: 공통 컴포넌트만 따로 분리하도록 폴더 구조 변경
